+++ src/telegram.h
@@ -212,7 +212,7 @@ class RxService : public EMSbus {
     void add(uint8_t * data, uint8_t length);
     void add_empty(const uint8_t src, const uint8_t dst, const uint16_t type_id);
 
-    uint32_t telegram_count() const {
+    uint16_t telegram_count() const {
         return telegram_count_;
     }
 
@@ -220,7 +220,7 @@ class RxService : public EMSbus {
         telegram_count_++;
     }
 
-    uint32_t telegram_error_count() const {
+    uint16_t telegram_error_count() const {
         return telegram_error_count_;
     }
 
@@ -256,8 +256,8 @@ class RxService : public EMSbus {
     static constexpr uint8_t EMS_BUS_QUALITY_RX_THRESHOLD = 5; // % threshold before reporting quality issues
 
     uint8_t                         rx_telegram_id_       = 0; // queue counter
-    uint32_t                        telegram_count_       = 0; // # Rx received
-    uint32_t                        telegram_error_count_ = 0; // # Rx CRC errors
+    uint16_t                        telegram_count_       = 0; // # Rx received
+    uint16_t                        telegram_error_count_ = 0; // # Rx CRC errors
     std::shared_ptr<const Telegram> rx_telegram;               // the incoming Rx telegram
     std::deque<QueuedRxTelegram>    rx_telegrams_;             // the Rx Queue
 };
@@ -305,7 +305,7 @@ class TxService : public EMSbus {
         return telegram_last_post_send_query_;
     }
 
-    uint32_t telegram_read_count() const {
+    uint16_t telegram_read_count() const {
         return telegram_read_count_;
     }
 
@@ -317,7 +317,7 @@ class TxService : public EMSbus {
         telegram_read_count_++;
     }
 
-    uint32_t telegram_fail_count() const {
+    uint16_t telegram_fail_count() const {
         return telegram_fail_count_;
     }
 
@@ -339,7 +339,7 @@ class TxService : public EMSbus {
         telegram_fail_count_++;
     }
 
-    uint32_t telegram_write_count() const {
+    uint16_t telegram_write_count() const {
         return telegram_write_count_;
     }
 
@@ -381,9 +381,9 @@ class TxService : public EMSbus {
   private:
     std::deque<QueuedTxTelegram> tx_telegrams_; // the Tx queue
 
-    uint32_t telegram_read_count_  = 0; // # Tx successful reads
-    uint32_t telegram_write_count_ = 0; // # Tx successful writes
-    uint32_t telegram_fail_count_  = 0; // # Tx unsuccessful transmits
+    uint16_t telegram_read_count_  = 0; // # Tx successful reads
+    uint16_t telegram_write_count_ = 0; // # Tx successful writes
+    uint16_t telegram_fail_count_  = 0; // # Tx unsuccessful transmits
 
     std::shared_ptr<Telegram> telegram_last_;
     uint16_t                  telegram_last_post_send_query_; // which type ID to query after a successful send, to read back the values just written
